Revert "Conforming to pep8 and folder restructure"
This reverts commit 7c51086df724bf8ccc58f840a24cbdfedd42776f.
