+++ mgmt/module.mk
@@ -1,4 +1,4 @@
-module := mizar
+module := mgmt
 
 submodules := tests
 -include $(patsubst %, $(module)/%/module.mk, $(submodules))
