+++ mgmt/tests/helper.py
@@ -4,7 +4,6 @@ logging.basicConfig()
 logger = logging.getLogger()
 logger.setLevel(logging.INFO)
 
-
 def do_add_remove_maglev(test, table, add, remove, exp_table, exp_prev_ele_map):
     logger.info("===Add/Remove Maglev test===")
     if add:
@@ -17,7 +16,5 @@ def do_add_remove_maglev(test, table, add, remove, exp_table, exp_prev_ele_map):
     act_prev_ele_map = table.get_prev_elements_map()
     test.assertEqual(exp_table, act_table)
     test.assertEqual(act_prev_ele_map, exp_prev_ele_map)
-    logger.info("Number of elements replaced {}".format(
-        table.elements_replaced))
-    logger.info("Number of elements replacing {}".format(
-        table.elements_replacing))
+    logger.info("Number of elements replaced {}".format(table.elements_replaced))
+    logger.info("Number of elements replacing {}".format(table.elements_replacing))
